+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 
 - Fixed lock of memory initialization in array creation
 - More platform-dependent function return error codes
+- Reduced the number of calls to `sll_platform_lock_acquire` and `sll_platform_lock_release`
 - Renamed `SLL_ERROR_FLAG_LIBC` and `error_codes$LIBC_ERROR_CODES` to `SLL_ERROR_FLAG_SYSTEM` and `error_codes$SYSTEM_ERROR_CODES`
 - Renamed argument parsing modifier symbolizing variadic functions from `+` to `!`
 - Rewritten the CLI executable loader code